Refactor - moved utils/ to parent folder
